+++ lib/src/lib.rs
@@ -9,6 +9,7 @@
 pub mod common;
 pub mod config;
 pub mod invidious;
+pub mod bilibili;
 pub mod new_database;
 pub mod player;
 pub mod playlist;