+++ tui/src/ui/components/popups/mod.rs
@@ -10,6 +10,7 @@ mod podcast;
 mod quit;
 mod saveplaylist;
 pub mod youtube_search;
+pub mod bilibili_search;
 
 #[allow(unused_imports)]
 pub use deleteconfirm::{DeleteConfirmInputPopup, DeleteConfirmRadioPopup};